added cost functions

+++ ParkingSpotsSearcher/src/model/University.cpp
@@ -4,7 +4,7 @@
 
 #include "University.h"
 
-University::University(long id, float latitude_degrees, float longitude_degrees) : Node(id, latitude_degrees,
+University::University(long id, double latitude_degrees, double longitude_degrees) : Node(id, latitude_degrees,
                                                                                         longitude_degrees) {}
 
 NodeType University::getType() const {
